refactor isolate ao autograd evaluation
