+++ tests/test_gpu4pyscf_ao_screening.py
@@ -26,14 +26,14 @@ except ModuleNotFoundError:
 
 from skala.functional.base import ExcFunctionalBase  # noqa: E402
 from skala.gpu4pyscf import SkalaKS  # noqa: E402
+from skala.pyscf.ao_evaluation import (  # noqa: E402
+    ChunkEvalForward,
+    non_chunk,
+)
 from skala.pyscf.backend import dft_gpu  # noqa: E402
 from skala.pyscf.evaluation import FeatureSpec  # noqa: E402
 from skala.pyscf.feature_math import MGGAFeatureFunction  # noqa: E402
-from skala.pyscf.features import (  # noqa: E402
-    ChunkEvalForward,
-    _prepare_spatially_sorted_grids,
-    non_chunk,
-)
+from skala.pyscf.features import _prepare_spatially_sorted_grids  # noqa: E402
 from skala.pyscf.numint import SkalaNumInt  # noqa: E402
 
 CARBON_CHAIN = """
